+++ src/cucu/init_data/features/steps/__init__.py
@@ -1,0 +1,7 @@
+# import all of the steps from cucu
+from cucu.steps import *  # noqa: F403, F401
+
+# import individual sub-modules here (i.e. module names of your custom step py files)
+# Example: For file features/steps/ui/login.py
+# import steps.ui.login_steps
+import steps.my_steps